requirements.txt
django==5.2
python-dotenv==1.0.0
google-genai
python-decouple==3.8
